requirements.txt
langchain
python-dotenv
langchain_community
langchain-text-splitters
langchain-openai
langchain_core
streamlit
langchain_groq
langchain-huggingface
pypdf